Refactoring passport logic, using MVC pattern #53
