+++ server/src/shared/index.ts
@@ -1,2 +1,3 @@
 export * from './Logger'
 export * from './redis'
+export * from './passport'
