+++ server/src/shared/redis.ts
@@ -1,6 +1,8 @@
+import _ from 'lodash'
 import session from 'express-session'
 import redis from 'redis'
 import redisStore from 'connect-redis'
+import { logger } from './Logger'
 
 let RedisStore = redisStore(session)
 let redisClient = redis.createClient()
@@ -9,8 +11,8 @@ let store = new RedisStore({ host: 'localhost', port: 6379, client: redisClient,
 export function getSessionUserId (sessionId: string) {
   return new Promise((resolve, reject) => {
     store.get(sessionId, (error, data) => {
-      if (error) {
-        reject(error)
+      if (error || !data) {
+        reject(error || `Redis cannot find user info for session`)
       } else if (data.passport.user.dbId) {
         resolve(data.passport.user.dbId)
       } else {
